Update new calculation logic for Service fulfillment and Enterprise Kpi

+++ backend/Models/EnterpriseKpiMetric.cs
@@ -30,9 +30,8 @@ namespace backend.Models
 
         public decimal? KpiValue { get; set; }
 
-        public byte Month { get; set; }
-
-        public short Year { get; set; }
+        public int Month { get; set; }
+        public int Year { get; set; }
 
         public DateTime CreatedAt { get; set; }
